+++ src/ui/index.ts
@@ -46,11 +46,17 @@ export type { CategorySelectProps } from "./CategorySelect";
 export { PrioritySelect } from "./PrioritySelect";
 export type { PrioritySelectProps, PrioritySelectVariant } from "./PrioritySelect";
 
+export { TypeSelect } from "./TypeSelect";
+export type { TypeSelectProps } from "./TypeSelect";
+
 export { DueDateBadge } from "./DueDateBadge";
 export type { BadgeVariant, DueDateBadgeProps } from "./DueDateBadge";
 
 export { CategoryBadge, CATEGORY_BADGE_FALLBACK } from "./CategoryBadge";
 export type { CategoryBadgeProps, CategoryBadgeVariant } from "./CategoryBadge";
+
+export { TypeBadge } from "./TypeBadge";
+export type { TypeBadgeProps, TypeBadgeVariant } from "./TypeBadge";
 
 export { Modal } from "./Modal";
 export type { ModalProps, ModalVariant } from "./Modal";